move TeamInfoTable component and tests from TeamDetail into their own files
