+++ hackathon_site/dashboard/frontend/src/pages/TeamDetail/TeamDetail.tsx
@@ -1,31 +1,24 @@
 import React from "react";
-import Header from "components/general/Header/Header";
-import Typography from "@material-ui/core/Typography";
+
 import TeamInfoTable from "components/teamDetail/TeamInfoTable/TeamInfoTable";
-import { mockTeamMultiple } from "testing/mockData";
-import {
-    Table,
-    TableBody,
-    TableContainer,
-    TableHead,
-    TableRow,
-    TableCell,
-    Checkbox,
-    Paper,
-    Container,
-    Grid,
-} from "@material-ui/core";
-import { RouteComponentProps } from "react-router-dom";
+
+import { RouteComponentProps, useHistory, useLocation } from "react-router-dom";
+import Header from "components/general/Header/Header";
 
 export interface PageParams {
     id: string;
 }
 
-const TeamDetail = () => {
+const TeamDetail = ({ match }: RouteComponentProps<PageParams>) => {
+    const history = useHistory();
+    const location = useLocation();
+
     return (
         <>
             <Header />
-            <TeamInfoTable match={id} />
+            <TeamInfoTable match={match} history={history} location={location} />
         </>
     );
 };
+
+export default TeamDetail;
